requirements.txt
Django==1.5.5
SQLAlchemy==0.8.3
South==0.8.2
Werkzeug==0.9.4
dj-database-url==0.2.2
djangorestframework==2.3.8
djorm-ext-pool==0.8
flake8==2.1.0
greenlet==0.4.0
gunicorn==18.0
ipython==1.1.0
mccabe==0.2.1
meinheld==0.5.5
mongoengine==0.8.4
pep8==1.4.6
psycopg2==2.5.1
pyflakes==0.7.3
pymongo==2.6.3
ujson==1.33
wsgiref==0.1.2
